+++ packages/index.ts
@@ -2,6 +2,7 @@ export type { FullScreenContainerProps } from './components/fullScreenContainer'
 export type { DigitalFlopProps } from './components/digitalFlop';
 export type { BorderBox1Props } from './components/borderBox1';
 export type { BorderBox2Props } from './components/borderBox2';
+export type { BorderBox3Props } from './components/borderBox3';
 export type { WaterLevelPondProps } from './components/waterLevelPond';
 export type { TabProps } from './components/tab';
 export type { ScrollRankingBoardProps } from './components/scrollRankingBoard';
@@ -11,6 +12,7 @@ export { default as FullScreenContainer } from './components/fullScreenContainer
 export { default as DigitalFlop } from './components/digitalFlop';
 export { default as BorderBox1 } from './components/borderBox1';
 export { default as BorderBox2 } from './components/borderBox2';
+export { default as BorderBox3 } from './components/borderBox3';
 export { default as WaterLevelPond } from './components/waterLevelPond';
 export { default as Tab } from './components/tab';
 export { default as ScrollRankingBoard } from './components/scrollRankingBoard';